added axios instance, moved api to config/
