--- wbooks/src/App/Auth/api.js
@@ -1,4 +1,0 @@
-export const API = {
-  baseURL: 'https://wbooks-api-stage.herokuapp.com/api/v1',
-  endpoint: '/users/sessions'
-}
