+++ wbooks/src/App/Home/Main/index.js
@@ -3,7 +3,7 @@ import { Switch, Route } from 'react-router-dom';
 import Dashboard from './Dashboard';
 import BookDetail from './BookDetail';
 import Profile from './Profile';
-import * as propTypes from '../../Constants/propTypes';
+import * as propTypes from '../../constants/propTypes';
 import './styles.css';
 
 class Main extends React.Component {
