+++ wbooks/src/App/constants/endpoints.js
@@ -1,0 +1,4 @@
+export const endpoints = {
+  login: 'users/sessions/',
+  signup: 'users/'
+}
